+++ modular_core/lib/src/di/injector.dart
@@ -28,7 +28,8 @@ class InjectorImpl<T> implements Injector<T> {
   }
 
   @mustCallSuper
-  bool isModuleAlive<T extends BindContext>() => _allBindContexts.containsKey(_getType<T>());
+  bool isModuleAlive<T extends BindContext>() =>
+      _allBindContexts.containsKey(_getType<T>());
 
   @mustCallSuper
   Future<bool> isModuleReady<M extends BindContext>() async {
@@ -44,7 +45,8 @@ class InjectorImpl<T> implements Injector<T> {
     final typeModule = module.runtimeType;
     if (!_allBindContexts.containsKey(typeModule)) {
       _allBindContexts[typeModule] = module;
-      (_allBindContexts[typeModule] as BindContextImpl).instantiateSingletonBinds(_getAllSingletons(), this);
+      (_allBindContexts[typeModule] as BindContextImpl)
+          .instantiateSingletonBinds(_getAllSingletons(), this);
       (_allBindContexts[typeModule] as BindContextImpl).tags.add(tag);
       debugPrint("-- $typeModule INITIALIZED");
     } else {
@@ -103,7 +105,9 @@ class InjectorImpl<T> implements Injector<T> {
 
   @override
   void updateBinds(BindContext context) {
-    final key = _allBindContexts.keys.firstWhere((key) => key.toString() == context.runtimeType.toString(), orElse: () => _KeyNotFound);
+    final key = _allBindContexts.keys.firstWhere(
+        (key) => key.toString() == context.runtimeType.toString(),
+        orElse: () => _KeyNotFound);
     if (key == _KeyNotFound) {
       return;
     }